requirements.txt
asgiref==3.8.1
Django==5.0.7
django-split-settings==1.3.2
psycopg==3.2.1
psycopg2==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
